Revert rename sock.h -> sock-impl.h
