+++ pkg/network/ebpf/c/prebuilt/tracer.c
@@ -25,7 +25,7 @@
 #include <uapi/linux/udp.h>
 #include <linux/err.h>
 
-#include "sock-impl.h"
+#include "sock.h"
 
 SEC("kprobe/tcp_sendmsg")
 int kprobe__tcp_sendmsg(struct pt_regs *ctx) {
